requirements.txt
requests==2.25.1
sshtunnel==0.4.0
python-dateutil==2.8.2
pandas>=1.0.0
# xorbits==0.3.2
# pyarrow==12.0.0
openpyxl==3.0.9
xlrd==1.2.0
# 中间件连接库
minio==7.1.1
sqlalchemy>=1.2.0
pymysql==1.0.2
elasticsearch>=7.15.2
redis==4.0.2
clickhouse-driver==0.2.3
clickhouse-sqlalchemy== 0.2.0
influxdb==5.3.1
py2neo==2020.0.0
kafka-python==2.0.2
mongoengine==0.24.2
mysql-replication==0.27